Refactor beforeEach function in test suites
