+++ tests/task.test.js
@@ -1,0 +1,4 @@
+const request = require("supertest");
+const Recipe = require("../src/models/recipe");
+
+test("Should create recipe for user", () => {});
